frontend/Extractor/gemini/requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
